Kalman filter

Kalman filter

Usage

kalman_filter(
  model,
  results_name,
  results_path = getwd(),
  kf_iterations_to_store = 2,
  write_to_file_at_each_iteration = TRUE,
  model_parameter_list = list(),
  algorithm_parameter_list = list(),
  external_packages = c(),
  julia_bin_dir = "",
  julia_required_libraries = c(),
  verify_cpp_function_types = FALSE,
  keep_temporary_model_code = FALSE
)

Arguments

model

A file containing the model, or a pre-compiled model list.

results_name

The name of the directory to which results will be written.

results_path

(optional) The path in which the results folder will be created (current working directory is the default).

kf_iterations_to_store

(optional) The number of iterations of filter output stored in memory as the algorithm is running (cannot be fewer than 2).

write_to_file_at_each_iteration

(optional) Do we write the algorithm output to file at each filtering step (TRUE/FALSE)?

model_parameter_list

(optional) A list containing parameters for the model.

algorithm_parameter_list

(optional) A list containing named parameters for the algorithm.

external_packages

(optional) A vector of names of other R packages the functions rely on.

julia_bin_dir

(optional) The directory containing the Julia bin file - only needed if Julia functions are used.

julia_required_libraries

(optional) Vector of strings, each of which is a Julia packge that will be installed and loaded.

verify_cpp_function_types

(optional) If TRUE, check the types of the parameters of user-defined .cpp functions. If FALSE (default), types are not checked.

keep_temporary_model_code

(optional) If FALSE (default), the .cpp file generated for compilation is deleted. If TRUE,

Value

Estimate of the marginal likelihood.